+++ Src/Library/Metadata.cs
@@ -1,5 +1,6 @@
 using System.Runtime.CompilerServices;
 
+[assembly: InternalsVisibleTo("FastEndpoints.AspVersioning")]
 [assembly: InternalsVisibleTo("FastEndpoints.Swagger")]
 [assembly: InternalsVisibleTo("FastEndpoints.Security")]
 [assembly: InternalsVisibleTo("FastEndpoints.UnitTests")]